--- src/api/Flooq.Api/Model/DataFlow.cs
@@ -1,9 +1,0 @@
-namespace Flooq.Api.Model
-{
-  public class DataFlow
-  {
-    public Guid Id { get; set; }
-    public string? Name { get; set; }
-    public string? Definition { get; set; }
-  }
-}
