+++ src/api/Flooq.Api/Models/DataFlow.cs
@@ -1,0 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Flooq.Api.Models
+{
+  public class DataFlow
+  {
+    [Key]
+    public Guid Id { get; set; }
+    public string? Name { get; set; }
+    public string? Status { get; set; }
+    public DateTime? LastEdited { get; set; }
+    public string? Definition { get; set; }
+  }
+}
